initialize pygame, add game loop, create gitignore
Initialized pygame and added the game loop - an infinite while loop. At
the moment, it simply draws a black screen. Also added an event
listener so that we can stop the program by clicking the window's close
button.

A note: I am using a universal import on constants.py, but only because
this is a small project, and the constants file is similarly small.
Otherwise, I would only import constants as needed.