add clock and delta time to limit to 60 FPS
I was a little confused by the instruction to call Clock.tick(60), then
divide it by 1000 and assign this value to dt. The .tick() method
pauses execution, but we save it to a variable; I think it's a method
that makes the change and also returns a value, like List.pop().