add asteroid.py and asteroidfield.py
The files create the asteroid and the asteroid field. The file
asteroidfield.py was written by boot.dev, but it looks like it defines
the edges of the screen, or, more likely the edges of the asteroid
field, then uses some random numbers and choices to set the size,
speed, and starting position of the asteroid.